Extract sleep timeout tables so the runtime suite tests production
The runtime state-machine tests defined their own copy of
getSleepTimeoutModeMs, so a wrong entry in the production table would
never fail a test. The labels, durations, and clamp now live in a pure
sleep_timeout_logic module that the test compiles directly; corrupting
a production entry fails the suite (verified). cyclefuncs keeps only
the cycle functions.
