+++ Firmware/src/cyclefuncs.cpp
@@ -12,27 +12,6 @@
 
 namespace
 {
-const char *SLEEP_TIMEOUT_MODE_LABELS[SLEEP_TIMEOUT_MODE_COUNT] = {
-    "Off",
-    "15s",
-    "30sec",
-    "1m",
-    "1m30s",
-    "2m"};
-
-const unsigned long SLEEP_TIMEOUT_MODE_MS[SLEEP_TIMEOUT_MODE_COUNT] = {
-    0,
-    15000,
-    30000,
-    60000,
-    90000,
-    120000};
-
-int clampSleepTimeoutMode(int timeout_mode)
-{
-  return constrain(timeout_mode, SLEEP_TIMEOUT_MODE_MIN, SLEEP_TIMEOUT_MODE_MAX);
-}
-
 int cycleValueWrapping(int current, int step, int minVal, int maxVal)
 {
   current += step;
@@ -324,15 +303,4 @@ void cycleLidarDistanceOffset()
   savePrefs(false, PREFS_DIRTY_SETTINGS);
 }
 
-const char *getSleepTimeoutModeLabel(int timeout_mode)
-{
-  int clampedMode = clampSleepTimeoutMode(timeout_mode);
-  return SLEEP_TIMEOUT_MODE_LABELS[clampedMode];
-}
-
-unsigned long getSleepTimeoutModeMs(int timeout_mode)
-{
-  int clampedMode = clampSleepTimeoutMode(timeout_mode);
-  return SLEEP_TIMEOUT_MODE_MS[clampedMode];
-}
 // ---------------------
